+++ Runtime/Settings/PopupSettings.cs
@@ -1,5 +1,4 @@
 using System;
-using UnityEngine;
 
 namespace AdvancedPS.Core.System
 {
@@ -9,8 +8,6 @@ namespace AdvancedPS.Core.System
         public InspectorEnum InspectorView { get; set; }
         public bool AutoSwitchInputModule { get; set; }
         public string LogType { get; set; }
-        public bool EscapeCloseEnabled { get; set; }
-        public KeyCode EscapeCloseKey { get; set; } = KeyCode.Escape;
     }
 
     public enum InspectorEnum : byte